refactored teacherUI
